Moved ignition_env.py
